+++ tests/test_pipeline_async.py
@@ -105,8 +105,21 @@ async def test_pipeline_nested_sync_trace(aconn, trace):
 
 async def test_cursor_stream(aconn):
     async with aconn.pipeline(), aconn.cursor() as cur:
-        with pytest.raises(psycopg.ProgrammingError):
-            await cur.stream("select 1").__anext__()
+        c1 = await aconn.execute("select 1")
+        assert not c1.pgresult
+        await cur.execute("select 2")  # results will be discarded by next stream()
+        series = [r async for r, in cur.stream("select generate_series(1, 4)")]
+        assert c1.pgresult
+    assert series == [1, 2, 3, 4]
+    assert (await c1.fetchone()) == (1,)
+
+
+async def test_cursor_stream_query_fetch_bug(aconn):
+    # See tests/pq/test_pipeline.py::test_pipeline_single_row_query_fetch_bug
+    # Worked around on Psycopg side by adding an extra sync().
+    async with aconn.pipeline(), aconn.cursor() as cur:
+        assert [r async for r, in cur.stream("select generate_series(0, 1)")] == [0, 1]
+        assert await (await aconn.execute("select 1")).fetchone() == (1,)
 
 
 async def test_server_cursor(aconn):